Extract 'resource endorsed event' to a shared example
