+++ decidim-core/lib/decidim/core/test.rb
@@ -71,3 +71,4 @@ require "decidim/core/test/shared_examples/categories_container_examples"
 require "decidim/core/test/shared_examples/assembly_announcements_examples"
 require "decidim/core/test/shared_examples/translated_event_examples"
 require "decidim/core/test/shared_examples/conversations_examples"
+require "decidim/core/test/shared_examples/resource_endorsed_event_examples"
